Reverted channel data type changes from previous commit, and set uart device as GNSS for testing purposes

+++ src/drivers/uart/Uart_public.h
@@ -88,8 +88,8 @@ typedef struct _Uart_Device {
     GPIO_PIN_INDEX gpio_pin_id_tx;
     uint8_t gpio_pin_rx;
     uint8_t gpio_pin_tx;
-    uint32_t udma_channel_tx;
-    uint32_t udma_channel_rx;
+    uint8_t udma_channel_tx;
+    uint8_t udma_channel_rx;
     Uart_Status uart_status_tx;
     Uart_Status uart_status_rx;
     Event uart_event;
